extract a bit more
