+++ packages/jest-runtime/src/index.ts
@@ -55,6 +55,7 @@ import {
   createOutsideJestVmPath,
   decodePossibleOutsideJestVmPath,
   findSiblingsWithFileExtension,
+  noop,
 } from './helpers';
 import CjsExportsCache from './internals/CjsExportsCache';
 import FileCache from './internals/FileCache';
@@ -82,6 +83,7 @@ import {
   buildJestGlobalsSyntheticModule,
   buildJsonSyntheticModule,
   buildWasmSyntheticModule,
+  evaluateSyntheticModule,
 } from './internals/syntheticBuilders';
 
 const esmIsAvailable = typeof SourceTextModule === 'function';
@@ -190,10 +192,6 @@ type WorklistEntry = {
   cacheKey: string;
   modulePath: string;
 };
-
-function noop() {
-  // empty
-}
 
 // `SourceTextModule#hasAsyncGraph()` lets us prove a graph is sync-evaluable.
 // `SyntheticModule` does not expose it but is by definition sync (the user
@@ -2864,41 +2862,4 @@ export default class Runtime {
   setGlobalsForRuntime(globals: JestGlobals): void {
     this.jestGlobals = globals;
   }
-}
-
-// On Node v22.21+ / v24.8+ a SyntheticModule starts in `'linked'` and the
-// body runs synchronously even though `evaluate()` returns a Promise -
-// return it sync so callers can store the actual module rather than a
-// Promise that can poison the registry if microtask draining stalls. On
-// older Node it starts `'unlinked'` and link/evaluate are genuinely async;
-// fall back to the async path there (the async-only legacy ESM code paths
-// handle the Promise return fine, and sync `require(esm)` doesn't exist on
-// those versions anyway).
-function evaluateSyntheticModule(
-  module: SyntheticModule,
-): SyntheticModule | Promise<SyntheticModule> {
-  if (module.status === 'unlinked') {
-    return evaluateSyntheticModuleAsync(module);
-  }
-  module.evaluate().catch(noop);
-  if (module.status === 'errored') {
-    throw module.error;
-  }
-  invariant(
-    module.status === 'evaluated',
-    `Synthetic module ${module.identifier} did not evaluate synchronously (status="${module.status}"). This is a bug in Jest, please report it!`,
-  );
-  return module;
-}
-
-async function evaluateSyntheticModuleAsync(
-  module: SyntheticModule,
-): Promise<SyntheticModule> {
-  await module.link(() => {
-    throw new Error('This should never happen');
-  });
-
-  await module.evaluate();
-
-  return module;
 }
